Move FailableStream functional interfaces to function package.
