+++ src/main/java/org/apache/commons/lang3/function/FailableBiFunction.java
@@ -17,26 +17,29 @@
 package org.apache.commons.lang3.function;
 
 /**
- * Represents a function that produces a long-valued result.  This is the
- * {@code double}-producing primitive specialization for {@link FailableFunction}.
+ * Represents a function that accepts two arguments and produces a result.
  *
- * <p>This is a functional interface whose functional method is {@link #applyAsDouble(Object)} (Object)}.
+ * <p>This is a functional interface whose functional method is {@link #apply(Object, Object)}.
  *
- * <p>An exception is thrown if an error occurs.
+ * <p>An exception will be thrown if an error occurs.
  *
- * @param <I> the type of the input to the function
- * @param <T> the type of exception to be thrown
+ * @param <I1> the type of the first argument to the operation
+ * @param <I2> the type of the second argument to the operation
+ * @param <O>  the type of the returned value
+ * @param <T>  the type of exception to be thrown
  *
- * @see java.util.function.ToDoubleFunction
+ * @see java.util.function.BiFunction
  * @since 3.10
  */
 @FunctionalInterface
-public interface FailableToDoubleFunction<I, T extends Throwable> {
+public interface FailableBiFunction<I1, I2, O, T extends Throwable> {
     /**
      * Apply the function.
-     * @param input the input for the function
+     *
+     * @param input1 the first input for the function
+     * @param input2 the second input for the function
      * @return the result of the function
      * @throws T if the function fails
      */
-    double applyAsDouble(I input) throws T;
+    O apply(I1 input1, I2 input2) throws T;
 }
