+++ src/main/java/org/apache/commons/lang3/function/FailablePredicate.java
@@ -17,26 +17,26 @@
 package org.apache.commons.lang3.function;
 
 /**
- * Represents a function that produces a long-valued result.  This is the
- * {@code double}-producing primitive specialization for {@link FailableFunction}.
+ * Represents a predicate (boolean-valued function) of one argument.
  *
- * <p>This is a functional interface whose functional method is {@link #applyAsDouble(Object)} (Object)}.
+ * <p>This is a functional interface whose functional method is {@link #test(Object)}.
  *
- * <p>An exception is thrown if an error occurs.
+ * <p>An exception will be thrown if an error occurs.
  *
  * @param <I> the type of the input to the function
  * @param <T> the type of exception to be thrown
  *
- * @see java.util.function.ToDoubleFunction
+ * @see java.util.function.Predicate
  * @since 3.10
  */
 @FunctionalInterface
-public interface FailableToDoubleFunction<I, T extends Throwable> {
+public interface FailablePredicate<I, T extends Throwable> {
     /**
-     * Apply the function.
-     * @param input the input for the function
-     * @return the result of the function
-     * @throws T if the function fails
+     * Test the predicate.
+     *
+     * @param object the object to test the predicate on
+     * @return the predicate's evaluation
+     * @throws T if the predicate fails
      */
-    double applyAsDouble(I input) throws T;
+    boolean test(I object) throws T;
 }
