+++ src/main/java/org/apache/commons/lang3/function/FailableToLongFunction.java
@@ -18,7 +18,7 @@ package org.apache.commons.lang3.function;
 
 /**
  * Represents a function that produces a long-valued result.  This is the
- * {@code long}-producing primitive specialization for {@link org.apache.commons.lang3.Functions.FailableFunction}.
+ * {@code long}-producing primitive specialization for {@link FailableFunction}.
  *
  * <p>This is a functional interface whose functional method is {@link #applyAsLong(Object)}.
  *
